refactor: organize components
remove lucia auth
