--- app.d.ts
@@ -1,9 +1,0 @@
-/// <reference types="lucia" />
-declare namespace Lucia {
-  type Auth = import("@/lib/auth/lucia").Auth;
-  type DatabaseUserAttributes = {
-    name: string;
-    email: string;
-  };
-  type DatabaseSessionAttributes = {};
-}
