+++ package.json
@@ -43,7 +43,6 @@
   },
   "dependencies": {
     "@hookform/resolvers": "^3.3.4",
-    "@lucia-auth/adapter-prisma": "^3.0.2",
     "@prisma/client": "^5.8.0",
     "@radix-ui/react-alert-dialog": "^1.0.5",
     "@radix-ui/react-dialog": "^1.0.5",
@@ -64,7 +63,6 @@
     "clsx": "^2.1.0",
     "dayjs": "^1.11.10",
     "jotai": "^2.6.1",
-    "lucia": "^2.7.6",
     "lucide-react": "^0.294.0",
     "next": "14.0.4",
     "react": "~18.2.0",
