+++ src/app/dashboard/(protected)/scan/page.tsx
