+++ src/components/dashboard/dashboard-navbar.tsx
@@ -1,10 +1,11 @@
 import Image from "next/image";
 import Link from "next/link";
 
+import { Button } from "@/components/ui/button";
+
 import { cn } from "@/lib/utils";
 
-import logoGTD from "../../public/static/images/logo-gtd-black-transparent.png";
-import { Button } from "./ui/button";
+import logoGTD from "../../../public/static/images/logo-gtd-black-transparent.png";
 
 export type DashboardNavbarProps = {
   className?: string;
