+++ src/components/registration/booking-item.tsx
