+++ src/components/registration/bundle-option.tsx
