+++ src/components/registration/bundle-select.tsx
