+++ src/components/registration/event-option.tsx
