+++ src/components/registration/registration-form.tsx
@@ -5,9 +5,6 @@ import { useAtom } from "jotai";
 import { useForm } from "react-hook-form";
 import { z } from "zod";
 
-import { formDataAtom } from "@/lib/atoms/events-registration";
-import { DEFAULT_REGISTRATION_FORM } from "@/lib/constants";
-
 import {
   Form,
   FormControl,
@@ -15,7 +12,10 @@ import {
   FormItem,
   FormLabel,
   FormMessage,
-} from "./ui/form";
+} from "@/components/ui/form";
+
+import { formDataAtom } from "@/lib/atoms/events-registration";
+import { DEFAULT_REGISTRATION_FORM } from "@/lib/constants";
 
 export const registrationFormSchema = z.object({
   name: z.string().min(3).max(50),
