+++ src/components/registration/stripe-form.tsx
