+++ src/components/registration/timeslot.tsx
