+++ src/components/registration/timeslots.tsx
