+++ src/lib/atoms/events-registration.ts
@@ -1,10 +1,10 @@
 import { atomWithStorage } from "jotai/utils";
 import { type z } from "zod";
 
-import { type registrationFormSchema } from "@/components/registration-form";
+import { type registrationFormSchema } from "@/components/registration/registration-form";
 
-import { DEFAULT_REGISTRATION_FORM } from "../constants";
-import { type LocalBooking } from "../types/events-registration";
+import { DEFAULT_REGISTRATION_FORM } from "@/lib/constants";
+import { type LocalBooking } from "@/lib/types/events-registration";
 
 const BASE_KEY = "event-registration";
