+++ src/lib/trpc/config.ts
@@ -1,13 +1,12 @@
 import { initTRPC } from "@trpc/server";
 import { ZodError } from "zod";
 
-import { auth } from "@/server/auth";
 import { db } from "@/server/db";
 
 import { transformer } from "./utils";
 
 export const createTRPCContext = async function (opts: { headers: Headers }) {
-  return { db, auth, ...opts };
+  return { db, ...opts };
 };
 
 const t = initTRPC.context<typeof createTRPCContext>().create({
